dme_controls: Use x86-64 indexers

+++ vgui2/dme_controls/dmedagrenderpanel.cpp
@@ -266,8 +266,8 @@ void CDmeDagRenderPanel::DrawJointNames( CDmeDag *pRoot, CDmeDag *pDag, const ma
 		g_pMatSystemSurface->DrawColoredText( m_hFont, vecPanelPos.x + 5, vecPanelPos.y, 255, 255, 255, 255, "%s", pJointName );
 	}
 
-	int nCount = pDag->GetChildCount();
-	for ( int i = 0; i < nCount; ++i )
+	intp nCount = pDag->GetChildCount();
+	for ( intp i = 0; i < nCount; ++i )
 	{
 		CDmeDag *pChild = pDag->GetChild(i);
 		if ( !pChild )
